simple_hubspot_MCP/requirements.txt
mcp-server>=0.0.1a2
hubspot-api-client>=8.1.0
python-dotenv>=1.0.1
requests>=2.31.0